requirements.txt
Django
djangorestframework
Pillow
python-decouple
monero
requests
channels
redis
pycryptodome
cryptography

